# openapi.yaml
...
openapi: "3.1.2"
info:
  title: Apache Nutch REST API
  description: >-
    REST API for managing Apache Nutch crawl jobs, configurations, seed lists,
    database queries, and data readers. The server is powered by Apache CXF
    with JAX-RS.
  version: "1.0.0"
  license:
    name: Apache 2.0
    identifier: Apache-2.0
  contact:
    name: Apache Nutch
    url: https://nutch.apache.org
servers:
  - url: "{protocol}://localhost:{port}"
    description: Nutch REST server
    variables:
      protocol:
        default: http
        enum:
          - http
          - https
        description: The protocol used to access the Nutch server.
      port:
        default: "8081"
        description: >-
          The port the Nutch server listens on. Configurable via the --port
          command-line argument.
security:
  - basicAuth: []
...
paths:
...
  /admin/:
    get:
      tags:
        - Admin
      summary: Get server status
      description: >-
        Returns the current status of the Nutch server including start date,
        known configurations, all jobs, and currently running jobs.
      operationId: getServerStatus
      responses:
        "200":
          description: Server status retrieved successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NutchServerInfo"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /admin/stop:
    get:
      tags:
        - Admin
      summary: Stop the Nutch server
      description: >-
        Initiates a graceful shutdown of the Nutch server. If jobs are still
        running and force is not set, the server will refuse to stop.
      operationId: stopServer
      parameters:
        - name: force
          in: query
          required: false
          description: >-
            If true, kills any running jobs before stopping the server.
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Shutdown status message.
          content:
            application/json:
              schema:
                type: string
              examples:
                stopping:
                  value: "Stopping in server on port 8081"
                busy:
                  value: "Jobs still running -- Cannot stop server now"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ---------------------------------------------------------------------------
  # Configuration
  # ---------------------------------------------------------------------------
  /config/:
    get:
      tags:
        - Configuration
      summary: List all configuration IDs
      description: Returns the set of all known configuration identifiers.
      operationId: getConfigs
      responses:
        "200":
          description: A JSON array of configuration ID strings.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                uniqueItems: true
              example:
                - default
                - my-custom-config
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /config/{configId}:
    get:
      tags:
        - Configuration
      summary: Get configuration properties
      description: Returns all key-value properties for the specified configuration.
      operationId: getConfig
      parameters:
        - $ref: "#/components/parameters/configId"
      responses:
        "200":
          description: A JSON object of configuration property key-value pairs.
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
              example:
                http.agent.name: "NutchBot"
                http.robots.agents: "NutchBot,*"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      tags:
        - Configuration
      summary: Delete a configuration
      description: >-
        Removes the specified configuration from the list of known
        configurations.
      operationId: deleteConfig
      parameters:
        - $ref: "#/components/parameters/configId"
      responses:
        "204":
          description: Configuration deleted successfully.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /config/{configId}/{propertyId}:
    get:
      tags:
        - Configuration
      summary: Get a single configuration property
      description: >-
        Returns the value of a specific property within the given
        configuration.
      operationId: getProperty
      parameters:
        - $ref: "#/components/parameters/configId"
        - $ref: "#/components/parameters/propertyId"
      responses:
        "200":
          description: The property value as plain text.
          content:
            text/plain:
              schema:
                type: string
              example: "NutchBot"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    put:
      tags:
        - Configuration
      summary: Update a configuration property
      description: >-
        Adds or updates the value of a property in the specified
        configuration.
      operationId: updateProperty
      parameters:
        - $ref: "#/components/parameters/configId"
        - $ref: "#/components/parameters/propertyId"
      requestBody:
        required: true
        description: The new property value as plain text.
        content:
          text/plain:
            schema:
              type: string
            example: "MyNewBot"
      responses:
        "200":
          description: Property updated successfully.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /config/create:
    post:
      tags:
        - Configuration
      summary: Create a new configuration
      description: >-
        Creates a new Nutch configuration with the specified parameters.
        Returns the configuration ID on success.
      operationId: createConfig
      requestBody:
        required: true
        description: The configuration to create.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NutchConfig"
      responses:
        "200":
          description: Configuration created. Returns the configuration ID.
          content:
            text/plain:
              schema:
                type: string
              example: "my-custom-config"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ---------------------------------------------------------------------------
  # Job
  # ---------------------------------------------------------------------------
  /job/:
    get:
      tags:
        - Job
      summary: List all jobs
      description: >-
        Returns job history for all jobs or filtered by crawl ID, regardless
        of job state.
      operationId: getJobs
      parameters:
        - name: crawlId
          in: query
          required: false
          description: Optional crawl ID to filter jobs by.
          schema:
            type: string
      responses:
        "200":
          description: A JSON array of job information objects.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JobInfo"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /job/{id}:
    get:
      tags:
        - Job
      summary: Get job info
      description: Returns detailed information for a specific job.
      operationId: getJobInfo
      parameters:
        - $ref: "#/components/parameters/jobId"
        - name: crawlId
          in: query
          required: false
          description: The crawl ID associated with the job.
          schema:
            type: string
      responses:
        "200":
          description: Job details.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobInfo"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /job/{id}/stop:
    get:
      tags:
        - Job
      summary: Stop a running job
      description: Attempts to gracefully stop a running job.
      operationId: stopJob
      parameters:
        - $ref: "#/components/parameters/jobId"
        - name: crawlId
          in: query
          required: false
          description: The crawl ID associated with the job.
          schema:
            type: string
      responses:
        "200":
          description: Whether the job was successfully stopped.
          content:
            application/json:
              schema:
                type: boolean
              example: true
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /job/{id}/abort:
    get:
      tags:
        - Job
      summary: Abort a job
      description: >-
        Forcefully aborts a job. Unlike stop, this kills the job immediately.
      operationId: abortJob
      parameters:
        - $ref: "#/components/parameters/jobId"
        - name: crawlId
          in: query
          required: false
          description: The crawl ID associated with the job.
          schema:
            type: string
      responses:
        "200":
          description: Whether the job was successfully aborted.
          content:
            application/json:
              schema:
                type: boolean
              example: true
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /job/create:
    post:
      tags:
        - Job
      summary: Create a new job
      description: >-
        Creates and enqueues a new Nutch job (e.g., inject, generate, fetch,
        parse, updatedb, index).
      operationId: createJob
      requestBody:
        required: true
        description: The job configuration specifying type, crawl ID, and arguments.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobConfig"
      responses:
        "200":
          description: Job created. Returns the job information.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobInfo"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ---------------------------------------------------------------------------
  # Database
  # ---------------------------------------------------------------------------
...
components:
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: HTTP Basic Authentication.

  # ---------------------------------------------------------------------------
  # Reusable Parameters
  # ---------------------------------------------------------------------------
  parameters:
    configId:
      name: configId
      in: path
      required: true
      description: The unique identifier for the configuration.
      schema:
        type: string
    propertyId:
      name: propertyId
      in: path
      required: true
      description: The name (key) of the configuration property.
      schema:
        type: string
    jobId:
      name: id
      in: path
      required: true
      description: The unique identifier for the job.
      schema:
        type: string
...
  schemas:
...
    NutchConfig:
      type: object
      description: Configuration for creating a new Nutch configuration.
      properties:
        configId:
          type: string
          description: The identifier for this configuration.
        force:
          type: boolean
          description: >-
            If true, overwrites an existing configuration with the same ID.
          default: false
        params:
          type: object
          additionalProperties:
            type: string
          description: Key-value pairs of Nutch configuration properties.
      example:
        configId: "my-config"
        force: false
        params:
          http.agent.name: "MyBot"
          http.robots.agents: "MyBot,*"
    JobConfig:
      type: object
      description: Configuration for creating a new crawl job.
      required:
        - type
      properties:
        crawlId:
          type: string
          description: The crawl identifier.
        type:
          $ref: "#/components/schemas/JobType"
        confId:
          type: string
          description: >-
            The configuration ID to use for this job. Defaults to "default"
            if not specified.
        jobClassName:
          type: string
          description: >-
            Fully qualified class name when type is CLASS.
        args:
          type: object
          additionalProperties: true
          description: Additional arguments for the job.
      example:
        crawlId: "crawl-01"
        type: "INJECT"
        confId: "default"
        args:
          seedDir: "seedFiles/seed-1700000000000"
...
    NutchServerInfo:
      type: object
      description: Status information about the running Nutch server.
      required:
        - configuration
        - jobs
        - runningJobs
      properties:
        startDate:
          type: string
          format: date-time
          description: The date and time the server was started.
        configuration:
          type: array
          items:
            type: string
          uniqueItems: true
          description: Set of known configuration IDs.
        jobs:
          type: array
          items:
            $ref: "#/components/schemas/JobInfo"
          description: All jobs (any state).
        runningJobs:
          type: array
          items:
            $ref: "#/components/schemas/JobInfo"
          description: Currently running jobs.
    JobInfo:
      type: object
      description: Information about a crawl job.
      required:
        - type
        - state
      properties:
        id:
          type: string
          description: The unique job identifier.
        type:
          $ref: "#/components/schemas/JobType"
        confId:
          type: string
          description: The configuration ID used for this job.
        args:
          type: object
          additionalProperties: true
          description: Arguments passed to the job.
        result:
          type: object
          additionalProperties: true
          description: Result data returned after job completion.
        state:
          $ref: "#/components/schemas/State"
        msg:
          type: string
          description: A human-readable status or error message.
        crawlId:
          type: string
          description: The crawl identifier associated with this job.
...
    JobType:
      type: string
      description: The type of Nutch crawl job.
      enum:
        - INJECT
        - GENERATE
        - FETCH
        - PARSE
        - UPDATEDB
        - INDEX
        - READDB
        - CLASS
        - INVERTLINKS
        - DEDUP
    State:
      type: string
      description: The current state of a job.
      enum:
        - IDLE
        - RUNNING
        - FINISHED
        - FAILED
        - KILLED
        - STOPPING
        - KILLING
        - ANY

  # ---------------------------------------------------------------------------
  # Reusable Responses
  # ---------------------------------------------------------------------------
  responses:
    BadRequest:
      description: >-
        Bad request. The request body is missing, malformed, or contains
        invalid parameters.
      content:
        text/plain:
          schema:
            type: string
          example: "Nutch configuration cannot be empty!"
    Unauthorized:
      description: >-
        Unauthorized. Basic authentication credentials are missing or
        invalid.
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
            example:
              message: "Authentication required."
    NotFound:
      description: The requested resource was not found.
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
            example:
              message: "Resource not found."
    InternalServerError:
      description: An unexpected server error occurred.
      content:
        text/plain:
          schema:
            type: string
          example: "Internal server error."